More myList work
I have been working on implementing myList. I have added a
myListIterator class which may or may not stay. However, I just had a
recent idea that may completely change the way this list functions.
Whether or not this idea is good depends on the way Java stores Objects
in Linked Lists; more research is required.